Added voice acting assets
Felt inspired in the moment so I made some voice acting support :D
I'll try to put the recordings into action tonight, if not, I'll do it over the weekend.
Not sure if they will mend good with the current theme & game feel, but if not, we will just leave them out.